Yeni klasör/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx[http2]==0.27.0
python-dotenv==1.0.1
pydantic==1.10.18


